requirements.txt
Django==2.1.3
django-filter==2.0.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
